backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
google-generativeai==0.3.2
python-multipart==0.0.6
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
aiohttp==3.9.1
Pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.1

